# contracts/openapi.yaml
...
openapi: 3.1.0
info:
  title: Bible Atlas API
  description: The Bible Atlas HTTP API, generated from the Rust that serves it.
  version: 0.8.0
paths:
...
  /api/chapter/{cref}:
    get:
      tags:
      - reading
      summary: |-
        `GET /api/chapter/{cref}`. `cref` must parse as exactly a
        `ScriptureRef::Chapter` (book + chapter, e.g. `EXO.14`) — a book-only or
        verse/passage-shaped path segment is the wrong shape for this endpoint
        and 400s as `bad_ref`, same as an unparseable one. The optional
        `?translation=kjv` query param (ruling 5, M1 is KJV-only) is never
        extracted, so its presence or absence cannot affect this handler at all.
      description: |-
        ruling-3-policy: once `cref` parses as a `Chapter`, an out-of-range
        chapter number (or a book with no known chapters in this atlas) is NOT an
        error — the verse-count bound comes from `canon.books[].chapters`, and an
        unknown/short chapter just yields `verse_count = 0`, i.e. a 200 response
        with an empty `verses` list. Same rationale as `scene_scripture`: a
        reader showing "no verses in this chapter" is a meaningful response, not
        a failure.

        Batch M-A (brief requirement 5, "re-implement the OLD /api/chapter
        handler as a VIEW over the window query"): the verse TEXT below now
        comes from `GraphState::chapter_span` + `GraphState::window` -- the SAME
        windowed reading-order query `GET /api/text?scope=chapter` calls --
        instead of `data.verses.get(key)`. The verse-count bound and the
        out-of-canon policy above are UNCHANGED and still sourced from
        `AtlasData`; headings moved to `graph.heading_index` in M-C2, and
        OVERLAY-1 Task 5 moved the PLACE-MENTION half onto the port too
        (`graph.scene_source(&data)`'s own `places_for_verse`/`place`, the
        materialised-from-the-graph successors of the deleted
        `AtlasData::places_for_verse`/`place_by_id` -- identical ids in
        identical order, see those methods' own doc comments). THIS endpoint
        (the reader's own chapter view) was untouched by Batch M-B's own
        event-world migration;
        only `/api/narrative/event/{id}` (see that handler's own doc comment)
        and the generic `/api/node`/`/edges` endpoints move to the graph this
        batch. The WIRE SHAPE is byte-for-byte identical -- proven by
        `tests/graph_equivalence.rs`'s own all-1,189-chapters comparison -- so
        every existing caller of this endpoint (the reader's chapter view/
        mini-reader/split view, `ChapterNode`, `PlaceCard`'s hover verse text,
        `PassageBlock`, `PopoverSectionProviders`) now serves from the graph
        with NO client-side change and no reader-visible behavior change.
      operationId: chapter
      parameters:
      - name: cref
        in: path
        required: true
        schema:
          type: string
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Chapter'
        '400':
          description: bad_window | bad_ref | bad_kind | bad_dir | bad_corpus
          content:
            application/json:
              schema:
                type: object
                required:
                - error
                properties:
                  error:
                    $ref: '#/components/schemas/ErrorInner'
                additionalProperties: false
        '404':
          description: not_found
          content:
            application/json:
              schema:
                type: object
                required:
                - error
                properties:
                  error:
                    $ref: '#/components/schemas/ErrorInner'
                additionalProperties: false
        '500':
          description: internal
          content:
            application/json:
              schema:
                type: object
                required:
                - error
                properties:
                  error:
                    $ref: '#/components/schemas/ErrorInner'
                additionalProperties: false
...
  /api/node/{id}:
    get:
      tags:
      - graph
      summary: |-
        `GET /api/node/{id}` (design doc §5): card (id/kind/label/provenance) +
        edge summary (kind -> true count, honesty needs it -- `GraphQuery`'s own
        `edge_summary` already lists only inhabited kinds) + the graph version
        stamp. `{id}` is the wire form `graph_wire::encode_node_id` produces; a
        malformed or unresolvable id is `bad_ref` (matches every other
        ref-shaped endpoint's own 400 convention), an id that parses but names
        no node in the built graph is `not_found` (six kinds resolve as of
        Batch M-B: `text-unit:BOOK.C.V` naming a real canon verse, and
        `Event:`/`Narrative:`/`Anchor:`/`Place:` naming a real curated/
        Theographic id; any other kind prefix, or a real-shaped id this batch's
        event world doesn't carry, 400s/404s per the same convention).
      operationId: node_card
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: string
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/NodeCard'
        '400':
          description: bad_window | bad_ref | bad_kind | bad_dir | bad_corpus
          content:
            application/json:
              schema:
                type: object
                required:
                - error
                properties:
                  error:
                    $ref: '#/components/schemas/ErrorInner'
                additionalProperties: false
        '404':
          description: not_found
          content:
            application/json:
              schema:
                type: object
                required:
                - error
                properties:
                  error:
                    $ref: '#/components/schemas/ErrorInner'
                additionalProperties: false
        '500':
          description: internal
          content:
            application/json:
              schema:
                type: object
                required:
                - error
                properties:
                  error:
                    $ref: '#/components/schemas/ErrorInner'
                additionalProperties: false
  /api/node/{id}/edges:
    get:
      tags:
      - graph
      summary: |-
        `GET /api/node/{id}/edges?kind=&cursor=&limit=` (design doc §5): one page
        of one edge kind. `kind` is a label from graph-types' own relation
        manifest (e.g. `"cites"`/`"cited-by"`) -- missing or unrecognized is
        `bad_kind`; `cursor` is the opaque (here: plain integer) offset the
        previous page's own `next` returned; `limit` defaults to 20, capped at
        200.
      operationId: node_edges
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: string
      - name: kind
        in: query
        required: true
        schema:
          $ref: '#/components/schemas/EdgeKind'
      - name: cursor
        in: query
        required: false
        schema:
          type: integer
          minimum: 0
      - name: limit
        in: query
        required: false
        schema:
          type: integer
          minimum: 0
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/EdgePage'
        '400':
          description: bad_window | bad_ref | bad_kind | bad_dir | bad_corpus
          content:
            application/json:
              schema:
                type: object
                required:
                - error
                properties:
                  error:
                    $ref: '#/components/schemas/ErrorInner'
                additionalProperties: false
        '404':
          description: not_found
          content:
            application/json:
              schema:
                type: object
                required:
                - error
                properties:
                  error:
                    $ref: '#/components/schemas/ErrorInner'
                additionalProperties: false
        '500':
          description: internal
          content:
            application/json:
              schema:
                type: object
                required:
                - error
                properties:
                  error:
                    $ref: '#/components/schemas/ErrorInner'
                additionalProperties: false
...
components:
  schemas:
...
    Chapter:
      type: object
      required:
      - ref
      - book
      - chapter
      - verses
      properties:
        book:
          type: string
        chapter:
          type: integer
          format: int32
          minimum: 0
        ref:
          type: string
        verses:
          type: array
          items:
            $ref: '#/components/schemas/Verse'
      additionalProperties: false
...
    EdgeKind:
      type: string
      enum:
      - contains
      - member-of
      - attested-in
      - attests
      - follows-in
      - precedes-in
      - dated-by
      - dates
      - located-at
      - site-of
      - mentions
      - mentioned-in
      - cites
      - cited-by
      - quotes
      - quoted-by
      - confesses
      - confessed-in
      - fulfilled-in
      - fulfills
      - prefigures
      - prefigured-by
      - named-after
      - namesake-of
      - justified-by
      - justifies
      - comments-on
      - commented-on-by
      - spoken-by
      - speech-of
      - spoken-at
      - site-of-speech
      - derived-from
      - derives
      - occurs-in
      - words
      - parent-of
      - child-of
      - participates-in
      - participants
      - analogous-to
      - catechism-link
      - corresponds-to
      - parallel
      - temporal-adjacency
      - partner-of
    EdgePage:
      type: object
      required:
      - kind
      - entries
      - version
      properties:
        entries:
          type: array
          items:
            $ref: '#/components/schemas/EdgeEntry'
        kind:
          $ref: '#/components/schemas/EdgeKind'
        next:
          type:
          - integer
          - 'null'
          minimum: 0
        version:
          type: string
      additionalProperties: false
    EdgeSummaryEntry:
      type: object
      required:
      - kind
      - count
      properties:
        count:
          type: integer
          minimum: 0
        kind:
          $ref: '#/components/schemas/EdgeKind'
      additionalProperties: false
...
    NodeCard:
      type: object
      required:
      - id
      - kind
      - label
      - provenance
      - edge_summary
      - version
      properties:
        description:
          type:
          - string
          - 'null'
          description: |-
            ENT-1a: Easton's Bible Dictionary (1897, PD) prose, source-attested,
            `None` until a match exists -- same additive-JSON, same held-client
            disclosure as `wire::PlaceDetail::description`. This is the
            ONLY "detail" surface a Person or PeopleGroup node has at all
            (`graph_wire.rs`'s own doc comment: no dedicated per-kind endpoint
            exists for either), so widening the generic card here is what
            actually reaches them; it reaches Place/PeopleGroup for free too
            (the same payload fact, whichever kind carries it).

            Batch CORP-1b: the SAME field, widened again -- a CommentaryItem's
            own prose (`NodePayload::CommentaryItem.text`) rides here too now
            (`atlas_graph::legacy::node_description`'s own updated match), for
            the identical reason: no dedicated per-kind endpoint exists for
            CommentaryItem either, and the prose was already sitting on the
            compiled graph payload, just never read.
        edge_summary:
          type: array
          items:
            $ref: '#/components/schemas/EdgeSummaryEntry'
        id:
          type: string
        kind:
          $ref: '#/components/schemas/NodeKind'
        label:
          type: string
        person:
          oneOf:
          - $ref: '#/components/schemas/PersonLife'
            description: 'D5: present for a Person only (omitted, never null, otherwise).'
          - type: 'null'
        provenance:
          type: string
        version:
          type: string
      additionalProperties: false
...